old-tests/submission/W3C/domhtml/HTMLFieldSetElement02.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - FieldSet</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/association-of-controls-and-forms.html#dom-fae-form" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<form action="..." id="form2" method="POST">
<fieldset>
<legend>All data entered must be valid</legend>
</fieldset>
</form>
<table summary="Table 1">
<tbody><tr>
<td>
<fieldset>
<legend>All data entered must be valid</legend>
</fieldset>
</td>
</tr>
</tbody></table>


<script>
var domtest = async_test("The form attribute returns null if control in not within the context of      form.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vform;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("fieldset");
      assert_equals(nodeList.length,2);
testNode = nodeList.item(1);
      vform = testNode.form;

      assert_equals(vform, null);
    
    domtest.done();
}));

</script>
<p>The form attribute returns null if control in not within the context of 
    form. 

    Retrieve the form attribute and examine its value.</p>
<div id="log"></div>
</body></html>
